requirements.txt
pyTelegramBotAPI
alembic
SQLAlchemy
psycopg2
loguru
python-dotenv
psycopg2-binary
pytest